Fixing issues when puppet runs as service
If puppet runs as service the following exceptions are raised:
The server threw an exception. (Exception from HRESULT: 0x80010105
(RPC_E_SERVERFAULT))
You cannot call a method on a null-valued expression.

Surrunded the method with try-catch and added better error handling
logic

Added a "virtual" provider and .NET based method, which is available
from windows .NET 4.5 (Windows 8) and made it default.

But left the original as a fallback.